Add support for osm import to use multiple rules.
If a mappings file contained more than one rule that
apply to a given feature, only the first of them was being
applied.
This patch makes sure that mapping with multiple rules to
multiple trees is possible on import, and that both
osm export-shp and osm export-sl can also export to multiple
tables given a mappings file with multiple rules.