Removes JLine2 as a dependency since it's not approved by LocationTech.
This patch removes the dependency on JLine2. Given there's no suitable
replacement we also get rid of the geogig-console command, which provided
an interactive shell to execute geogig commands.

We do still depend on JAnsi, for colored output to the terminal.

A new class of our own cooking, org.locationtech.geogig.cli.Console,
replaces the use of JLine's ConsoleReader. This new class is minimalistic
and attends only to the functionality required by geogig-cli, in terms
of writing to STDOUT or the provided output stream.