Allow to limit the number of reported diff directly at PreOrderDiffWalk.
Added a boolean return value to PreorderDiffWalk.Consumer.feature(...) to allow
for quiclky aborting the whole diff tree traversal.

It is different than the return value of Consumer.tree(...) and Consumer.bucket(...)
in that those only forbid going deeper into the trees they're given as argument.

The new PreorderDiffWalk.MaxFeatureDiffsConsumer Consumer implementation allows to
abort the whole tree traversal once the given limit has been reached.

This feature can also be used in other scenarios where the whole diffs are not
to be traversed, like in finding a tree child or aborting the traversal at any
point (say because of a running action that's cancelled).

Added a limit argument to the DiffTree operation accordingly and made Status
make use of it so we don't end up with non-fully consumed iterators in the
status report.