Import osm way.nodes as a long[] and way.tags as a Map<String, String>.
Instead of storing way.nodes and way.tags as strings, store them
as long[] and Map<String, String> instead.

The osm export commands functionality of exporting nodes and tags as Strings
is preserved, given the target GeoTools FeatureStores do not support long[]
and Map properties natively.

Conversion to and from String is delegated to the GeoTools Converters
framework (hence working transparently) by means of the
PrimitiveArrayToStringConverterFactory and MapToStringConverterFactory
converter factories.

All tests pass, including the ones that use mappings, though were adapted
to expect long[] or Map<String, String> accordingly.

The patch already provides a measurable improvement both in time and space.
An osm import takes a bit less time, and a rather small one like importing
the latest antarctica pbf provides for a 100MB reduction in the JE db size
(i.e. results in a 1.6 instead of 1.7 GB db). Yet the space savings are
expected to be bigger once the V3 serialization format is ready, as it uses
delta compression and variable lenght integer encoding for ints and longs.