Support serialization of large character strings.
So far large(ish) strings couldn't be serialized, as DataOutput.writeUTF()
has a max (byte encoded) length of 65535.

The worse case scenario is that the UTF-8 representation of all characters in
the string is 3 bytes.

This patch splits strings larger than 65535/3 into chunks of length 65535
and encodes them as consecutive UTF-8 strings.

The 2-byte UTF-8 length header is abused using 65535 as a special value
indicating the string is split into chunks, and if so an int with the
number of chunks is written. This is safe since given the above splitting
rule we'll never actually write a string that results in 65535 bytes when encoded.