Reverting "Random" repository directory name.
Commit b7a0b615483c8e55669b9bac8a02143a33930ac5 did a few other things
to fix some tests due to repository directories being "shared" across
multiple tests. I'm still not sure what the issue was, but this random
string appended to the repository in the test makes the repo directory
name "unpredictable" for other tests using the RepositoryTestCase.

I discovered this in trying to build the GeoGig GeoServer Extension. One
of the test in that project was failing because it was trying to verify
that the repository still existed after the GeoGig store is shutdown. But
it fails becuase it expects the repo directory to be named "repo", not
"repo-<random number>". Since the directory isn't exposed in the test
class here, external tests using it can't reliably predict what the
repo directory name will be and can't verify it exists by name.

I'm planning on submitting a fix on geoserver-exts for the GeoGig plugin,
but it will obviously depend on this fix getting merged in. The line in
particular is here:
https://github.com/boundlessgeo/geoserver-exts/blob/master/geogig/src/test/java/org/geogig/geoserver/wfs/WFSIntegrationTest.java#L380