Optimize RevTree structure to avoid an overwhelming exponential growth in the number of bucket trees.
Up to this commit, the splitting of RevTree objects into subtrees was mandated by
the RevTree.MAX_BUCKETS (32) and RevTree.NORMALIZED_SIZE_LIMIT (512) constants.

Meaning that the maximum capacity of a leaf tree object before having to be split
into subtrees was always 512, and the number of subtrees an overflowed leaf tree
was split into was always 32, as shown in the first table bellow.

This lead to an exponential growth in the number of internal trees as powers of 32
as the RevTree depth level increases.

While this presented little trouble for datasets with less than 2^24 features (roughly
16 million) at depth level 2 (i.e. third depth level), which resulted in 32768 leaf trees,
the next depth level results in 2^20 (>1M) leaf trees for over 500 million features, and
the next one in over 33 million leaf trees for up to 17 billion features, and so forth.

Now, those start being pretty big leaps, when it comes to handle that number of objects
in commodity hardware and default JVM settings. Despite the max numbers are surely reached
when the max number of features are too, the number of leaf trees tend to be that big
even when the total number of features is far from the max capacity, due to the way feature
nodes are sorted and assigned to a bucket (by using a hash function on the feature identifier
with a high distribution uniformity), which is one of the most essential characteristics of
geogig since it allows to compute differences between trees really fast.

This commit defines both the number of bucket trees a leaf tree is split into, and
the maximum capacity of a leaf tree, as functions of the tree depth, by replacing
the RevTree.MAX_BUCKETS constant by the
NodePathStorageOrder.maxBucketsForLevel(int depthIndex):int function, and the
RevTree.NORMALIZED_SIZE_LIMIT by the NodePathStorageOrder.normalizedSizeLimit(int depthIndex):int
function, as shown in the second table bellow.

NOTE the settings for the first three levels of nesting are left intact for backwards
compatibility with existing repositories. Rationale being that the projects currently
using GeoGig (which hasn't yet reached its first final release) aren't using datasets
bigger than 16 million features, and 32K leaf trees with 1056 bucket trees are easily
handled.

Past the third level of nesting, the buckets per node is decreased to avoid the explosion
in number of leaf trees, and the leaf tree max capacity of a leaf tree reduced from 512
to 256 to also decrease the storage size needs. The result is a canonical tree structure
that increases more gradually in storage and capacity, with more realistic limits (i.e.
up to 17 billion features at the eight depth level instead of 562 trillions (which if ever
reached would require an estimated storage capacity of over 24000 terabytes to hold the
leaf trees, and over 40 terabytes to hold the internal bucket trees that lead to the leaf
trees [1]).

===================================================================================
RevTree structure before this commit:
===================================================================================
Depth #buckets	      Leaf trees	Total # of	#features    Max #features
      per node                  	tree objects    per leaf
                                                	tree
-----------------------------------------------------------------------------------
0	32	               32	           33	512	             16,384
1	32	            1,024	        1,056	512	            524,288
2	32	           32,768	       33,792	512	         16,777,216
3	32	        1,048,576	    1,081,344	512	        536,870,912
4	32	       33,554,432	   34,603,008	512	     17,179,869,184
5	32	    1,073,741,824	1,107,296,256	512	    549,755,813,888
6	32	   34,359,738,368      35,433,480,192	512	 17,592,186,044,416
7	32	1,099,511,627,776   1,133,871,366,144	512	562,949,953,421,312
===================================================================================

==========================================================================
RevTree structure after this commit:
==========================================================================
Depth #buckets	  Leaf trees	 Total # of     #features    Max #features
      per node                   tree objects   per leaf
                                                tree
--------------------------------------------------------------------------
0	32	        32	         33	512	            16,384
1	32	     1,024	      1,056	512	           524,288
2	32	    32,768	     33,792	512	        16,777,216
3	 8	   262,144	    294,912	256	        67,108,864
4	 8	 2,097,152	  2,359,296	256	       536,870,912
5	 4	 8,388,608	 10,485,760	256	     2,147,483,648
6	 4	33,554,432	 41,943,040	256	     8,589,934,592
7	 2	67,108,864	100,663,296	256	    17,179,869,184
==========================================================================

[1] For an estimated average Node object size of 47 byes and bucket node size
    of 40 bytes, both in uncompressed form .