Make PostOrderDiffWalk thread safe
Commit 3b726f18b46ff81d02d59ec1f09920dc2cd0b5eb, which
makes PreOrderDiffWalk run on a ForkJoinPool (multiple
threads) introduced a bug that prevented fetch operations
from a remote repo from succeeding in certain situations.

The root cause of it was that PostOrderDiffWalk, which
wraps PreOrderDiffWalk, wasn't make thread safe, and it
used a simple Stack to know when to report back inner trees.

This commit fixes PostOrderDiffWalk to support its Consumer
being called from concurrent threads.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>