Configure default user when creating an empty repo in WEB API functional tests.
The 'Import to an empty repository' geopackage scenario
was succeeding in the development box because there
was a /home/groldan/.geogigconfig file with user.name and user.email
configured, and failing in the CI server because there's no
such file in the jenkins' home dir.

This patch makes it so WebAPICucumberHooks.setUpEmptyRepo()
also calls <repo>/config? to set user.name and user.email
to the repo's local config.