Initial import of geogig-ql (Query Language) prototype
The geogig-ql module is an experimental module to support
a SQL like DDL syntax for handling data.

It currently has rudimentary support for SELECT, INSERT,
UPDATE, and DELETE.

The module is part of the "experimental" maven profile,
so it's only built and included as part of the CLI app
if the profile is enabled.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>