Improve performance for removal of conflicts in ConflictsDatabase.
The only way to remove a subset of conflicts from ConflictsDatabase
was to delete them one by one. This patch adds the method
removeConflicts(@Nullable String namespace, Iterable<String> paths)
and makes it so Index.stage() don't need to remove potential conflicts
one by one.
It also introduces org.locationtech.geogig.storage.PersistedIterable
as a helper class to avoid OOM issues by saving objects to a temporary
file that can then be used as an iterable (in this case as argument
to ConflictsDatabase.removeConflicts()).

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>