Fix a bug when merging Feautres and Geometry attributes
- The use of Object.equals() for JTS Geometry objects in
GeometryAttributeDiff resulted in calling Geometry.equalsTopo()
which produced TopologyException's. Replaced by calls
to Geometry.equalsExact() which compares coordinate by
coordinate instead of a topological comparison, whcih is
what was originally intended.

- GeometryAttributeDiff.conflics(AttributeDiff) was not considering
the case where the geometry diff represents a no change, so it
returned true (conflict) for cases where it wasn't so.

Also on this patch:
- MergeFeaturesOp is simplified and easier to read. It also fetches
all needed objects from the ObjectDatabase at once instead of one
by one, which when inside a loop should provide for a performance
improvement.

- AttributeDiff (and its realizations for generic and geometric
attributes) was abusing the use of Optional. Optional was being
used both as arguments, return values, and instance variables,
and with weird semantics, where Optional arguments were allowed
to be null. Got rid of all the Optionals in these classes and
marked what's can be null explicitly, resulting in a more compact
and clear code.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>