IMPORTANT BUG EXPOSED: leaf RevTree node storage order is wrong.
Add a test to PreorderDiffWalk that demonstrates that traversing
a diff does not produce the feature events in the expected
order.

Leaf RevTrees are storing Nodes in an incorrect order, which
leads to parallel comparison of tree diffs for merge to return
Node iterators in inconsistent order.

The underlying reason is NodePathStorageOrder, the Comparator that
represents the canonical ordering of nodes for trees, having a
compare(String, String):int implementation that's not symmetric
to the buckets assigned to each feature id. The net result is
Tree traversal is wrong at the leaf tree node level.

The fix forcefully will produce BACKWARDS INCOMPATIBLE RevTree
structures, so any repository created before 1.0-RC3 is basically broken.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>