Make GeometryAttributeDiff report conflict when both branches changed the geometry
GeometryAttributeDiff was being too smart, performing
a merge of geometries changed by both sides of the merge
which could result in topologically invalid geometries.

Besides, it's imposible to know what the user would do
when both branches changed the same geometry, so better
just mark a conflict and let the user resolve it.

GeometryAttributeDiff can still be used to apply a patch
though.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>