Enforce normalization of Polygon geometries
This patch makes sure GeoGig always stores polygon
geometries in normalized form (that is, outer ring
coordinates in CW order, inner rings in CCW order).

This way, an import/export/re-import cycle won't
produce unwanted geometry changes if the original
polygons were illformed.

The bug was encountered when importing a polygons
shapefile, exporting to another shapefile,
and re-importing, it is possible that several
features get changed on the geometry attribute,
when none should have.

This also means GeoGig stores polygons in normalized
form so if someone is crazy enough as to encode
information in the orientation of polygon rings
then they need to rethink what they're doing.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>