Fix Incompatible Feature Type error for equivalent CRSs
During import, minor differences between CRSs cause an Incompatible Feature Type error.
Inside GeoGIG, the CRS is stored along the lines of "EPSG:4326".  However, when it is
compared to an imported Shapefile, the check is too "exact" and it detects a CRS
difference when there isn't really a difference.

This fix makes the check of CRS equivelency less strict.

Signed-off-by: DBlasby <dblasby@boundlessgeo.com>