Fix (c) header for org.locationtech.geogig.storage.postgresql.EnvironmentBuilderTest
The GPL header was mistakenly introduced when the file was
created due to having the Eclipse code template for GeoServer
activated instead of the GeoGig code template while Erik
was working on both projects at the same time.

This patch replaces it by the correct (c) header used by this project.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>