Add NOTICE file with dependencies
This commit added a NOTICE file and a template from which the
NOTICE file is created. Licenses which are not found online are
added to the 'licenseMapping.xml' file.

Current NOTICE file is modified (by hand) to not include any
dependencies which aren't from the 'compile' scope.
Re-executing the 'mvn notice:generate' command will replace the current
NOTICE file with a new one.

See issue: GIG-451

Signed-off-by: goudine <agoudine@boundlessgeo.com>