Refactor API packages in core module as *.impl
Packages in the core module with matching names in
the api module are preventing proper jar signing.

Since all classes and interfaces in the api module
(packages model, repository, and storage) are fully
public, we decided to refactor the matching packages
in the core module as model.impl, repository.impl, and
storage.impl.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>