Add method RevTreeBuilder.update(Node old, Node new)
update(old, new) is an opportunity for RevTreeBuilder
implementations to optimize for the case when a node
is being replaced instead of just added.

In any case the result must be the same as calling
remove(oldNode) followed by put(newNode), and both
oldNode and newNode must refer to the same node name.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>