release resources of ForkJoin pool
There are two ForkJoin pools being used by geogig
this makes this a bit more explicit.  There is a SHARED_FORKJOIN_POOL
that is used when making unordered requests (preserveOrder=false),
and then a private (one per request) single-thread when order needs
to be preserved.

At the end of the Walk, the private pool is released.

Signed-off-by: dblasby <dblasby@boundlessgeo.com>