Reduce number of db roundtrips when using a FeatureSource/Reader
Adds experimental Context.snapshot():Context method to use
a view of the repository at that point in time, by cacheing
all the values in the RefDatabase.

This reduces the number of roundtrips to the RefsDatabase
for a single GeoServer GetMap request from 21 to just 1, which
is especially cumbersome for the postgres backend on slow-ish
networks.