allows FeatureReaderBuilder to return property-reduced features
The geogig FeatureReaderBuilder currently always returns all the properties in the feature - even if the underlying Query only asks for a few of the properties.  This change just retypes the resulting features so they only have the requested properties.

If you don't do this, the WFS doesn't correctly return features when a request comes in for a property-reduces Query/GetFeatures.

@groldan -- not sure if this is the correct way to do this given the lazyness of the features.  It might be possible to do retyping at an earlier stage (i.e. instead of creating the full features in the first place).

Signed-off-by: dblasby <dblasby@boundlessgeo.com>