Fix a bug that made updates to spatial index trees contain the old and new nodes
QuadTreeClusteringStrategy.buildRoot() collapses top level DAG's
that are composed of one single bucket to the shallowest child DAG
that's either a leaf tree or has more than one bucket.

This is so many nodes that fall on the same quadrant do not produce
unnecessary deep trees.

When such a RevTree representing a spatial index is updated (that is,
a new one is created based on a previous one), the original structure
needs to be recreated (the top level one-bucket DAGs) so that node updates
can subsequently find their path to the tree structure and be correctly
updated.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>