Create FeatureSources that respect the provided namespace and definition query
Commit db4da25 introduced a deffect by which the namespce URI
assigned to a DataStore would not be propagated to the SimpleFeatures
(although it would be propagated to the FeatureSource's FeatureType),
due to FeatureBuilder receiving the RevFeatureType instead of a
SimpleFetureType to avoid the misleading perception that it could
build a Feature in any FeatureType given instead of being forced
to match an actual geogig RevFeatureType.

This patch fixes the returned feature's target name/namespace
and also makes it so GeoGigFeatureSource supports the "definition
query" properties that might be given to its ContentFeatureSource
superclass.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>