Quad trees: add maxDepth, round quad bounds to float-32, add pre/post condition checks
This is an intermediate commit for the (unfinished)
ability of quad-trees to collapse single-quad DAG paths
before building the final RevTree index, and symmetrically
expand DAG's when creating a new index based on a previous one.

Since collapse/expand is unfinished but apparently on the right
track, this commit disables collapse/expand through the
QuadTreeClusteringStrategy.ENABLE_EXPAND_COLLAPSE constant, in
order to still take advantage of:

- Quadrand.slice() rounding to float-32 value space to match
the rounding applied to org.locationtech.geogig.model.Node.bounds(),
so that quad addressing is consistent with Node's value range.
Otherwise the expansion applied to Node bounds would lead to
either a node not reaching it's expected max depth. The natural
order of the Quadrant enum makes sure the slight overlap that
might occur between sibling quadrants will never lead to
inconsistencies in determining the quadrant for a node.

- maxDepth: the maximin depth for a quad tree is not computed
from it's maxBounds, as the deepest depth at which a point node
which has been expanded in both directions due to float-32
value space rounding can fully fit on the quadrant's bounds.
This way, edge case sets of data (like a point cluster in a very
small area) would not produce unnecessarily deep index trees,
but will defer to a canonical (so called unpromotables in index
jargon) as a direct child of the deepest index level.

- pre/post condition checks: a branch ref update after a commit,
merge, rebase, etc. make the built in UpdateIndexesHook command hook
to update the indexes for the feature trees that have changed,
ultimately calling BuildIndexOp for each of them. BuildIndexOp
now makes sure the resulting index is consistent with the
canonical feature tree by checking on the return codes of
RevTreeBuilder.put()/remove()/update() as apporpriate, and ensuring
the resuling index tree has the same size than the canonical
feature tree.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>