Fix a bug in shriking and re-expanding DAG's in ClusteringStrategy
ClusteringStrategy.shrinkIfUnderflow() didn't reset the bucket DAGs
to the empty tree, producing trees with the wrong number of nodes
when nodes where removed and then new nodes were added forcing
the bucket node to re-expand. It nows clears any DAG that had
an original RevTreeId so the old tree nodes are not rea-added.

This patch also removes unnecessary arguments to shrinkIfUnderflow()
and mergeRoot().

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>