Fix stack overflow in RevParse
On a repo with a rather large number of commits,
trying to log an old enough ancestor produced a
StackOverflowException at RevParse.resolveAncestor(...).

Changed it to avoid recursion and speed it up considerably
by querying the repository's GraphDatabase to reach out
to the Nth parent instead of parsing each and every
actual RevCommit.

Another not-so-small speed optimization is avoiding as
many calls to RevObjectParse as possible and instead
access the ObjectStore directly.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>