Implementation for a shared RevObject cache
A RevObject cache to be transparently shared among repository
instances.

A repository (or ObjectStore implementation) supplies a unique
identifier to CacheManager.acquire(String) to obtain an
ObjectCache that, from the point of view of the calling code,
contains only the RevObject instances it puts on it, but that
internally shares the cache with other ObjectCache instances,
so that the cache size can be tracked and managed as a single
entity for all.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>