Fix empty string bug in custom geotools converter factories
Fix a bug in PrimitiveArrayToStringConverterFactory that
returned an empty array when converting an empty string to
a primitive array instead of null.

It had the side effect of breaking the geoserver config UI
for ImageMosaic layers as it uses the Converters framework
to parse the "Bands" argument to an int[] that comes as an
empty string.

This patch makes the same fix to our MapToStringConverterFactory
although no bug report was issued for it, for the sake of
consistency.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>