Avoid opening the repo to check for the ansi.enabled config option
ansi.enabled shall be a global config option. Before this patch,
CLI was opening the repo in read-write mode just to get to fetch
that config setting, also making read-only commands fail on
repositories with the rocksdb backend that were open by another
process.

This patch resolves the repository URI, if given as an argument or
resolvable from the current working directory, and then uses
RepositoryResolver.resolveConfigDatabase(...) to get the (possibly
global only) config database instead, avoiding the attempt to open
all the repository's databases in read-write mode and then closing
it, just to be re-opened later to actually execute the command.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>