Avoid opening the repo just to unalias a CLI command
In case a CLI command can't be found, a check is performed
by GeogigCLI to verify if the command name is an alias in
the config database. To do so, it was opening the repository
and closing it (meaning connecting to all it's databases),
to then opening and closing it again to actually execute
the command.

This patch opens only the ConfigDatabase to check for the alias.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>