Avoid failure to return the pg config db if the repository does not exist.
PGRepositoryDatabase.getConfigDatabase(...) assumed the repository pointed out
by the URI exists if the repository name is present in the URI, in order to set
the storage.refs, storage.objects, etc config properties default values (which
is odd by the way, and will be not necessary when we simplify the plugins
mechanism).

Yet, the repository may not exist and the config db still may be needed to set/get
global options. This caused an exception, now prevented by checking if the repo
exists before trying to set the storage config settings.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>