Make CLI get the ansi.enabled config from file based global config only
The ansi.enabled config option only makes sense on a per terminal/OS/machine
basis and hence we shall only use it through the $HOME/.geogigconfig
global config. This also avoids errors trying to get the config value
from the PG repository resolver which is not friendly with the rootUri
boolean argument of its getConfigDatabase mehtod.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>