Move http remote classes to web-api module
Breaks an implicit dependency from core to web-api.
The web-api module depends on core, and defines the
HTTP interface for interacting with remotes, yet the
client side of the fence was on core.

This patch moves the client side of HTTP remoting
to web-api so that the module is self contained.

To that end, a new java.util.ServiceLoader extension point,
org.locationtech.geogig.remote.RemoteResolver, has been
added in order to being able of resolving the IRemoteRepo
at runtime based on the Remote configuration.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>