Fix a bug by which re-hashing a commit would result in a different SHA-1
HashObjectFunnels.CommitFunnel was funneling the commit id,
when the id is not part of the contents of a RevObject, and
hence it shall not be considered when computing it's hash.

For backwards compatibility with versions prior to 1.2,
CommitFunnel now has a hardcoded ObjectId.NULL null id
funneled where the commit ids was being funneled before.

This workaround is ok because the only actual way RevCommit
instances were created is through CommitBuilder, which first
created a fake RevCommit with NULL id, hashed it out, and used
the resulting id to create the final RevCommit instance.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>