Ensure test case commit times are unique + logOp doesn't re-traverse
This was originally manifest as a windows build failure.  Root cause was that the log times
were exactly the same time - resulting in the LogOp returning rows in the wrong order.

This fixes;
a) LogOp - fixed implementation so it doesn't accidently follow a history path twice
b) TestPlatform - ensure that times are always increase (never equal)