Temporary meassure to fix the build on Windows for the core module
Ensure test case commit times are unique + logOp doesn't re-traverse

This was originally manifested as a windows build failure.  Root cause was that the log times
were exactly the same time - resulting in the LogOp returning rows in the wrong order.

This fixes:
a) LogOp - fixed implementation so it doesn't accidently follow a history path twice
b) TestPlatform - ensure that times are always increase (never equal)

Nonetheless, this is just a temporary fix, the real fix needs to:
- Add explicit test cases for when the commits have the same timestamp
- Make sure the historiracl log order is predictable in those cases