Single PG connection pool per test class
Running the pg tests implied the creation and destruction
of several connection pools per test case, for the same database.

This patch introduces PGTestDataSourceProvider as a JUnit
ExternalResource to be used either as an @Rule or @ClassRule,
so that the javax.sql.DataSource can be used on a test class or
test case basis, as appropriate.

Most test classes are updated to configure the test DataSource
once for all its test cases.