Storage API change: make GraphDatabase owned by ObjectDatabase
* Make GraphDatabase returned by ObjectDatabase.getGraphDatabase().
* Move StorageType from geogig-api to geogig-core. It's a storage
implementation detail and doesn't belong on geogig-api.
* Delete FileObjectDatabase, it's dead code.
* Bind Refs, Objects, and Index databases to Providers.of(null) in
the default GeogigModule Guice module, forcing concrete modules to
provide actual implementations. ConfigDatabase is still bound to
IniFileConfigDatabase to contraint this patch to just the API change,
but in the future it needs to be bound to null too, and provide a
semi-persistent implementation of HeapConfigDatabase in MemoryModule to
aid in unit testing without having to create repository directories.
* Move TestConfigDatabase from geogig-api test classes to geogig-core's
HeapConfigDatabase, and add conformance tests for it.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>