Transfer responsibility for updating graph to ObjectDatabase
Removes the decorator used to update the GraphDatabase whenever
a RevCommit was added to the ObjectDatabase, transferring that
responsibility to the ObjectDatabase itself, which owns the
GraphDatabase.

This makes for a huge performance gain when dealing with lots of commits

The Postgres ObjectDatbse can also use the same connection/transaction
being used to insert the objects accounting for extra consistency
guarantees between the revision objects and the commit graph.

Also, the PG backend can now figure out which version the server is
running (see the new Version class and PGStorage.getServerVersion()
method in geogig-postgres), which is used to make the geogig_graph_edge
table inserts using the new UPSERT functionality if the server runs PG
9.5+, or a slightly less efficient approach otherwise.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>