Several housekeeping changes to remoting related code
* Code readability: CloneOp and FetchOp reorganized for simplicity and
readability.
* IRemoteRepo: made it so it doesn't require a hold to the local
repository, but instead it's given to the methods that require it as
argument.
* RemoteRepositoryTestCase's way of preparing remotes was too cumbersome
to follow with too many Mockito.spy'ed objects just to mock up the
resolution of the IRemoteRepo instances for fetch,clone,lsremote,push,
and pull. Added a new OpenRemote command instead, and mocked it up only
once at the test's set up.
* Introduce CommandFactory as a superinterface to hold the single
<T extends AbstractGeoGigOp<?>> T command(Class<T> commandClass) method,
and made IRemoteRepo implement it with a default stub method as the
basis to execute selected commands on remotes as RPC calls.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>