Fix a bug in PreorderDiffWalk by which some bucket events wouldn't be notified
The diff traversal was not calling PreorderDiffWalk.Consumer.bucket(...)
when comparing a bucket tree against a leaf tree at either side of the
comparison. Instead, it was deferring directly to the feature() or
tree() consumer callbacks as it traversed down the bucket children
against the applicable nodes from the leaf tree at the other side of the
comparison.

This, although correct to compute Node diffs (i.e. feature and tree
nodes), had the side effect that it was impossible for a consumer to be
reported of all the bucket trees.

This patch makes sure Consumer.bucket() is called for every bucket
regardless of wehther there's a matching bucket at the same index and
depth, and adds the actual RevTree instances being compared as payload
to the BucketIndex Consumer.bucket(...) argument so that a consumer can
know whether the event is triggered because at either side of the
comparison there's a leaf tree or not, and make decisions based on that.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>